+++ Memory/MemoryHeader.h
@@ -8,7 +8,7 @@ class MemoryHeader
 {
 public:
 	MemoryHeader(uint32 allocSize);
-	static auto AttachHeader(MemoryHeader* header, size_t allocSize) -> void*;
+	static auto AttachHeader(MemoryHeader* header, uint32 allocSize) -> void*;
 	static auto DetachHeader(void* ptr) -> MemoryHeader*;
 	auto GetAllocSize() -> uint32;
 	auto SetAllocSize(uint32 allocSize) -> void;
